// shopping-list-app/src/styles/interfaces/IStylePadding.ts
export default interface IStylePadding {
  //
  // padding all sides
  //
  p1?: boolean;
  p2?: boolean;
  p3?: boolean;
  p4?: boolean;
  p5?: boolean;
  //
  // padding left
  //
  pl1?: boolean;
  pl2?: boolean;
  pl3?: boolean;
  pl4?: boolean;
  pl5?: boolean;
  //
  // padding right
  //
  pr1?: boolean;
  pr2?: boolean;
  pr3?: boolean;
  pr4?: boolean;
  pr5?: boolean;
  //
  // padding top
  //
  pt1?: boolean;
  pt2?: boolean;
  pt3?: boolean;
  pt4?: boolean;
  pt5?: boolean;
  //
  // padding bottom
  //
  pb1?: boolean;
  pb2?: boolean;
  pb3?: boolean;
  pb4?: boolean;
  pb5?: boolean;
}